feat: Implement Repayment model and update Transaction amount to `int`
- Add `Repayment` data model with fields: `id`, `transactionId`, `amount`, `when`, `createdAt`.
- Configure `Repayment` model for JSON serialization with snake_case naming.
- Implement validation to ensure repayment amount is positive and the date is not in the future.
- Add unit tests for the `Repayment` model, covering instantiation, validation, and JSON serialization.
- Change the `amount` field type from `double` to `int` in the `Transaction` model and update related tests to use integer values for currency.
- Add implementation plan for TASK-0009 and mark it as complete in the backlog.

+++ lib/domain/models/transaction.dart
@@ -10,7 +10,7 @@ class Transaction {
   final String id;
   final TransactionType type;
   final String name;
-  final double amount;
+  final int amount;
   final Currency currency;
   final String? description;
   final DateTime? dueDate;
@@ -56,7 +56,7 @@ class Transaction {
     String? id,
     TransactionType? type,
     String? name,
-    double? amount,
+    int? amount,
     Currency? currency,
     String? description,
     DateTime? dueDate,
